endorsement: use types with clear width

+++ src/syscalls.c
@@ -1290,9 +1290,9 @@ bolos_err_t os_pki_get_info(uint8_t                  *key_usage,
 
 bolos_err_t sys_endorsement_get_code_hash(uint8_t *hash, size_t length)
 {
-    unsigned int parameters[2] = {0};
+    uint32_t parameters[2] = {0};
     parameters[0]          = (uintptr_t) hash;
-    parameters[1]              = (unsigned int) length;
+    parameters[1]          = length;
     return (bolos_err_t) SVC_Call(SYSCALL_ENDORSEMENT_GET_CODE_HASH_ID, parameters);
 }
 
@@ -1300,8 +1300,8 @@ bolos_err_t sys_endorsement_get_public_key(ENDORSEMENT_slot_t slot,
                                            uint8_t           *public_key,
                                            size_t            *public_key_length)
 {
-    unsigned int parameters[3] = {0};
-    parameters[0]              = (unsigned int) slot;
+    uint32_t parameters[3] = {0};
+    parameters[0]          = slot;
     parameters[1]          = (uintptr_t) public_key;
     parameters[2]          = (uintptr_t) public_key_length;
     return (bolos_err_t) SVC_Call(SYSCALL_ENDORSEMENT_GET_PUB_KEY_ID, parameters);
@@ -1311,8 +1311,8 @@ bolos_err_t sys_endorsement_get_public_key_signature(ENDORSEMENT_slot_t slot,
                                                      uint8_t           *sig,
                                                      size_t            *sig_length)
 {
-    unsigned int parameters[3] = {0};
-    parameters[0]              = (unsigned int) slot;
+    uint32_t parameters[3] = {0};
+    parameters[0]          = slot;
     parameters[1]          = (uintptr_t) sig;
     parameters[2]          = (uintptr_t) sig_length;
     return (bolos_err_t) SVC_Call(SYSCALL_ENDORSEMENT_GET_PUB_KEY_SIG_ID, parameters);
@@ -1320,9 +1320,9 @@ bolos_err_t sys_endorsement_get_public_key_signature(ENDORSEMENT_slot_t slot,
 
 bolos_err_t sys_endorsement_key1_get_app_secret(uint8_t *secret, size_t length)
 {
-    unsigned int parameters[2] = {0};
+    uint32_t parameters[2] = {0};
     parameters[0]          = (uintptr_t) secret;
-    parameters[1]              = (unsigned int) length;
+    parameters[1]          = length;
     return (bolos_err_t) SVC_Call(SYSCALL_ENDORSEMENT_KEY1_GET_APP_SECRET_ID, parameters);
 }
 
@@ -1331,9 +1331,9 @@ bolos_err_t sys_endorsement_key1_sign_data(const uint8_t *data,
                                            uint8_t       *sig,
                                            size_t        *sig_length)
 {
-    unsigned int parameters[4] = {0};
+    uint32_t parameters[4] = {0};
     parameters[0]          = (uintptr_t) data;
-    parameters[1]              = data_length;
+    parameters[1]          = data_length;
     parameters[2]          = (uintptr_t) sig;
     parameters[3]          = (uintptr_t) sig_length;
     return (bolos_err_t) SVC_Call(SYSCALL_ENDORSEMENT_KEY1_SIGN_DATA_ID, parameters);
@@ -1344,9 +1344,9 @@ bolos_err_t sys_endorsement_key1_sign_without_code_hash(const uint8_t *data,
                                                         uint8_t       *sig,
                                                         size_t        *sig_length)
 {
-    unsigned int parameters[4] = {0};
+    uint32_t parameters[4] = {0};
     parameters[0]          = (uintptr_t) data;
-    parameters[1]              = (unsigned int) data_length;
+    parameters[1]          = data_length;
     parameters[2]          = (uintptr_t) sig;
     parameters[3]          = (uintptr_t) sig_length;
     return (bolos_err_t) SVC_Call(SYSCALL_ENDORSEMENT_KEY1_SIGN_WITHOUT_CODE_HASH_ID, parameters);
@@ -1357,9 +1357,9 @@ bolos_err_t sys_endorsement_key2_derive_and_sign_data(const uint8_t *data,
                                                       uint8_t       *sig,
                                                       size_t        *sig_length)
 {
-    unsigned int parameters[4] = {0};
+    uint32_t parameters[4] = {0};
     parameters[0]          = (uintptr_t) data;
-    parameters[1]              = (unsigned int) data_length;
+    parameters[1]          = data_length;
     parameters[2]          = (uintptr_t) sig;
     parameters[3]          = (uintptr_t) sig_length;
     return (bolos_err_t) SVC_Call(SYSCALL_ENDORSEMENT_KEY2_DERIVE_AND_SIGN_DATA_ID, parameters);
@@ -1369,7 +1369,7 @@ bolos_err_t sys_endorsement_get_metadata(ENDORSEMENT_slot_t slot,
                                          uint8_t           *metadata,
                                          size_t            *metadata_length)
 {
-    unsigned int parameters[3] = {0};
+    uint32_t parameters[3] = {0};
     parameters[0]          = slot;
     parameters[1]          = (uintptr_t) metadata;
     parameters[2]          = (uintptr_t) metadata_length;
